refactor(stable_api): extract TradingMixin into pyquotex._api.trading
Moves 7 trade-execution methods (buy, sell_option, open_pending,
check_win, get_result, get_profit, get_history) from the monolithic
stable_api.py into a focused mixin. Quotex inherits from TradingMixin;
public surface unchanged (verified by tests/test_api_surface.py).
